test: move gitUtils and ProxiedFS tests to `tests` folder
